fix: Separate TrailMapButton and Modal to avoid duplicated rendering
