+++ src/components/Menu/DesktopMenu.tsx
@@ -4,7 +4,7 @@ import * as Styled from './styled'
 import * as CommonStyled from '../../styles/styled'
 import { Event } from '../../generated/dreamkast-api.generated'
 import Link from 'next/link'
-import { TrailMapModal } from '../TrailMap/TrailMapModal'
+import { TrailMapButton } from '../TrailMap/TrailMapButton'
 
 type Props = {
   event?: Event
@@ -14,13 +14,13 @@ type Props = {
 export const DesktopMenu: React.FC<Props> = ({ event, url }) => {
   return (
     <Styled.DesktopMenu>
-      <TrailMapModal />
       <Link href={`/${event?.abbr}/ui`} rel="noreferrer">
         <Button style={{ color: '#423A57' }}>Top</Button>
       </Link>
       <Link href={`/${event?.abbr}/ui/info`} rel="noreferrer">
         <Button style={{ color: '#423A57' }}>Info</Button>
       </Link>
+      <TrailMapButton />
       <CommonStyled.MenuLink href={`/${event?.abbr}/o11y`} rel="noreferrer">
         <Button style={{ color: '#423A57' }}>Grafana</Button>
       </CommonStyled.MenuLink>
