+++ src/components/TrailMap/TrailMapModal.tsx
@@ -1,40 +1,23 @@
-import React, { useState } from 'react'
-import { Badge, Modal } from '@material-ui/core'
+import React from 'react'
+import { Modal } from '@material-ui/core'
 import { StampCard } from './StampCard'
-import * as Styled from './styled'
-import { stampSelector } from '../../store/settings'
-import { useSelector } from 'react-redux'
+import { settingsSelector, setTrailMapOpen } from '../../store/settings'
+import { useDispatch, useSelector } from 'react-redux'
 
 type Props = {
   todo?: boolean
 }
 
 export const TrailMapModal = (_: Props) => {
-  const [open, setOpen] = useState<boolean>(false)
-  const stamp = useSelector(stampSelector)
+  const settings = useSelector(settingsSelector)
+  const dispatch = useDispatch()
   return (
-    <>
-      <Badge
-        badgeContent={'!'}
-        color="secondary"
-        invisible={!stamp.canGetNewStamp}
-      >
-        <Styled.TrailMapButton
-          onClick={() => setOpen(true)}
-          className={stamp.canGetNewStamp ? 'pulse' : ''}
-        >
-          Trail Map
-        </Styled.TrailMapButton>
-      </Badge>
-      <Modal
-        open={open}
-        onClose={() => setOpen(false)}
-        aria-labelledby="modal-modal-title"
-        aria-describedby="modal-modal-description"
-      >
-        {/*TODO*/}
-        <StampCard></StampCard>
-      </Modal>
-    </>
+    <Modal
+      open={settings.isTrailMapOpen}
+      onClose={() => dispatch(setTrailMapOpen(false))}
+    >
+      {/*TODO*/}
+      <StampCard></StampCard>
+    </Modal>
   )
 }
